move util function to tokio_util.rs
