+++ src/ops.rs
@@ -41,6 +41,7 @@ use tokio::net::TcpListener;
 use tokio::net::TcpStream;
 use tokio_process::CommandExt;
 use tokio_threadpool;
+use tokio_util;
 
 type OpResult = DenoResult<Buf>;
 
@@ -468,34 +469,7 @@ where
   if is_sync {
     Box::new(futures::future::result(f()))
   } else {
-    Box::new(poll_fn(move || convert_blocking(f)))
-  }
-}
-
-/// `futures::future::poll_fn` only support `F: FnMut()->Poll<T, E>`
-/// However, we require that `F: FnOnce()->Poll<T, E>`.
-/// Therefore, we created our version of `poll_fn`.
-fn poll_fn<T, E, F>(f: F) -> PollFn<F>
-where
-  F: FnOnce() -> Poll<T, E>,
-{
-  PollFn { inner: Some(f) }
-}
-
-struct PollFn<F> {
-  inner: Option<F>,
-}
-
-impl<T, E, F> Future for PollFn<F>
-where
-  F: FnOnce() -> Poll<T, E>,
-{
-  type Item = T;
-  type Error = E;
-
-  fn poll(&mut self) -> Poll<T, E> {
-    let f = self.inner.take().expect("Inner fn has been taken.");
-    f()
+    Box::new(tokio_util::poll_fn(move || convert_blocking(f)))
   }
 }
